app/requirements.txt
flask
requests
boto3
python-keycloak
flask_cors
